Restructuring Project
Models, views and forms placed in folders
