+++ home/forms/articleForm.py
@@ -1,0 +1,7 @@
+from django import forms
+from ..models.articleModel import Article
+
+class ArticleForm(forms.ModelForm):
+    class Meta:
+        model = Article
+        fields = '__all__'
