+++ home/forms/authForm.py
@@ -1,5 +1,4 @@
 from django import forms
-from .models import Article
 
 class RegisterForm(forms.Form):
     email = forms.EmailField(label='Email', max_length=100)
@@ -10,8 +9,3 @@ class RegisterForm(forms.Form):
 class LoginForm(RegisterForm):
     email = None
     password2 = None
-
-class ArticleForm(forms.ModelForm):
-    class Meta:
-        model = Article
-        fields = '__all__'
